allow only  authorized users to edit and delete posts

+++ controllers/feed.js
@@ -141,6 +141,13 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error
             }
+            //check if creator id is equal to the id of the currently logged in user
+            //id of token received
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
 
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
@@ -177,6 +184,13 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error
         }
+            //check if creator id is equal to the id of the currently logged in user
+            //id of token received
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             //check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(post)
